% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMode.R
\name{myMode}
\alias{myMode}
\title{Mode}
\usage{
myMode(x)
}
\arguments{
\item{x}{R object}
}
\value{
The mode of \code{x}.
}
\description{
This function allows you to calculate the mode.
}
\examples{
myNumbers <- c(7,6,6,10,6,6,7,3,7,14,20,7)
myMode(myNumbers)
}
\keyword{mode}
