% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMultipleMode.R
\name{myMultipleMode}
\alias{myMultipleMode}
\title{Mode}
\usage{
myMultipleMode(values, multiple = FALSE)
}
\arguments{
\item{multiple}{TRUE will retur multiple variables, FALSE will show only one of the modes. Defaults to FALSE.}

\item{x}{R object}
}
\value{
A list with the mode of \code{x} and the times it was repeated.
}
\description{
This function allows you to calculate multiples mode if your data is multimodal.
}
\examples{
myNumbers <- c(7,6,6,10,6,6,7,3,7,14,20,7)
myMultipleMode(myNumbers)
myMultipleMode(myNumbers, TRUE)
}
\keyword{mode}
