% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlotMode.R
\name{myPlotMode}
\alias{myPlotMode}
\title{Plot numerical data highlighting the mode}
\usage{
myPlotMode(x, ...)
}
\arguments{
\item{x}{R object}
}
\description{
This function allows you to do a boxplot of your data and draw a vertical line where the mode is located
}
\examples{
myNumbers <- c(7,6,6,10,6,6,7,3,7,14,20,7)
myPlotMode(myNumbers, col = "blue")
}
\keyword{mode}
